# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QLGD API",
        "description": "Tài liệu OpenAPI cho hệ thống Quản lý Giảng dạy.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8888",
            "description": "Máy chủ API"
        }
    ],
    "paths": {
...
        "/api/lecturer/makeup-requests": {
            "get": {
                "tags": [
                    "Lecturer - Dạy bù"
                ],
                "summary": "Danh sách đề xuất dạy bù",
                "operationId": "lecturerMakeupIndex",
                "responses": {
                    "200": {
                        "description": "Danh sách phân trang",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MakeupRequestResource"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lecturer - Dạy bù"
                ],
                "summary": "Đề xuất buổi dạy bù",
                "operationId": "lecturerMakeupStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "leave_request_id",
                                    "suggested_date",
                                    "timeslot_id"
                                ],
                                "properties": {
                                    "leave_request_id": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "suggested_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-28"
                                    },
                                    "timeslot_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "example": 12,
                                        "nullable": true
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Ưu tiên phòng lab",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo đề xuất dạy bù thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MakeupRequestResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy đơn nghỉ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lecturer/makeup-requests/{id}": {
            "get": {
                "tags": [
                    "Lecturer - Dạy bù"
                ],
                "summary": "Chi tiết đề xuất dạy bù",
                "operationId": "lecturerMakeupShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 200
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết đề xuất",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MakeupRequestResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lecturer - Dạy bù"
                ],
                "summary": "Hủy đề xuất dạy bù",
                "operationId": "lecturerMakeupDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 200
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Đã hủy"
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Chỉ hủy khi PENDING",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Lecturer - Dạy bù"
                ],
                "summary": "Chỉnh sửa đề xuất dạy bù (khi còn PENDING)",
                "operationId": "lecturerMakeupUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 200
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "suggested_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-30"
                                    },
                                    "timeslot_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "example": 9,
                                        "nullable": true
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Đề xuất phòng đa năng",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MakeupRequestResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Chỉ chỉnh sửa khi PENDING",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Thông tin xác thực không hợp lệ."
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 120
                    }
                },
                "type": "object"
            },
...
            "MakeupRequestResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 200
                    },
                    "leave_request_id": {
                        "type": "integer",
                        "example": 100
                    },
                    "suggested_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-28"
                    },
                    "timeslot_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "room_id": {
                        "type": "integer",
                        "example": 8,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "PENDING"
                    },
                    "note": {
                        "type": "string",
                        "nullable": true
                    },
                    "decided_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "decided_by": {
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter the access token returned after login (omit the Bearer prefix)."
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}